Uploading OptiM Linac Files
This is all the work so far on the NL and SL for the weakly focusing triplet option. Final versions are also uploaded to the sharepoint. The "final" versions can be located in the "Updated" subfolder within each linac folder (example: /NL/Updated/) The periodic sections are just *L_trip_#.opt and the full linacs are *L_trip_#_Extended.opt.