Splitter Geometry
The splitter geometry for 6 passes is completed for a first iteration. This is for the NE corner, but extraction dipoles have been used as placeholders so both sides can have the same geometry.

The 4 necessary files are NESplitter_test.bmad, tao_NESplit.startup, tao_splitters.init, and Wall_Splitters.in

To run this, have all 4 files in the same directory and run: tao -init tao_splitters.init

The next iterations will make minor corrections, as well as start addressing path length.

Finally, please note that these are not cleaned up files. There are parts commented out, and the naming will have to change later when cleaned up.