Stewart CEBAF example
Adding some example files sent by Stewart Boogert (Cockcroft). These should run with 1.5 mm beam pipe, and the universe was made "vacuum" instead of air for visualization. The proper energy range cuts are included. This is ARC1 only - and the energy is not set to the real energy for troubleshooting.